Move go packages to pkg
