+++ pkg/api/status.go
@@ -6,7 +6,8 @@ import (
 
 	"github.com/go-chi/chi"
 	"github.com/go-chi/render"
-	"github.com/stefanprodan/mgob/db"
+
+	"github.com/stefanprodan/mgob/pkg/db"
 )
 
 type appStatus []*db.Status
