+++ pkg/notifier/smtp.go
@@ -6,7 +6,8 @@ import (
 	"strings"
 
 	"github.com/pkg/errors"
-	"github.com/stefanprodan/mgob/config"
+
+	"github.com/stefanprodan/mgob/pkg/config"
 )
 
 func sendEmailNotification(subject string, body string, config *config.SMTP) error {
